+++ source/CHANGELOG.md
@@ -4,14 +4,14 @@ Andreas Abel <andreas.abel@gu.se>  (unreleased)
 
 ## Major improvements
 
-* Haskell: layout keywords can now be stacked on the same line [#354]
-  See https://bnfc.readthedocs.io/en/latest/lbnf.html#stacking-layout-keywords
+* Haskell: layout keywords can now be stacked on the same line [#354],
+  see https://bnfc.readthedocs.io/en/latest/lbnf.html#stacking-layout-keywords
 * C/C++ backends now create reentrant parsers [#349]
 
 ## Bug fixes and small improvements
 
 * Haskell-GADT: generated code is warning free [#346]
-* Haskell: fixes in layout preprocessor [#343,#344,#345,#352,#353]
+* Haskell: fixes in layout preprocessor [#343,#344,#345,#352,#353,#356]
 * C: fixed a space leak when parsing from a string in memory [#347]
 * C: new methods `free*` to deallocate syntax trees [#348]
 * C: removed errorneous `define`d constructors from `Skeleton.c`